+++ src/lib.rs
@@ -3,4 +3,5 @@ pub mod data;
 pub mod domain;
 pub mod parser;
 pub mod service;
+pub mod trash;
 pub mod ui;